Add Generics to Service  file to handle different API requests
