+++ AppStoreProject/Controllers/New Group/AppsHorizontalController.swift
